requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
blis==0.7.11
breadability==0.1.20
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
deptry==0.16.1
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.111.0
fastapi-cli==0.0.4
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
logging==0.4.9.6
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.0
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.6
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pipdeptree==2.23.0
preshed==3.0.9
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==1.7.1
pymongo==4.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
replicate==0.27.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scipy==1.14.0
setuptools==70.2.0
shellingham==1.5.4
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
